fix fetching data after 1h
